backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pandas>=2.2.2
scikit-learn>=1.5.0
python-dotenv>=1.0.1
google-genai>=0.1.0
regex>=2023.12.25
pydantic>=2.8.2